Remove unused variable
This variable is shadowed by the declaration 'var last' in the for loop below, so it actually ends up unused. There isn't really a reason to declare it up here unless you need to use it within the loop as well as after the loop. It's generally best to keep related things as close together as possible, when possible, and that includes the declaration of a variable and the usage of it