Use console.error when logging errors
This is mostly just a good habit. In node, there's not much different between log and error, primarily that console.log writes to stdout and console.error writes to stderr, but that can be important if scripting around the output at all. On the frontend, browser consoles will display console.error in red (background included) and will include a stack trace automatically, so it has more immediate value